+++ injectable_generator/lib/import_resolver.dart
@@ -15,27 +15,32 @@ class ImportResolverImpl extends ImportResolver {
 
   String resolve(Element element) {
     // return early if source is null or element is a core type
-    if (element.source == null || _isCoreDartType(element)) {
+    if (element?.source == null || _isCoreDartType(element)) {
       return null;
     }
 
     for (var lib in libs) {
-      if (lib.source != null && !_isCoreDartType(lib) && lib.exportNamespace.definedNames.keys.contains(element.name)) {
+      if (lib.source != null &&
+          !_isCoreDartType(lib) &&
+          lib.exportNamespace.definedNames.keys.contains(element.name)) {
         return lib.source.uri.toString();
       }
     }
     return null;
   }
 
-  static String relative(String packagePath, String from) {
-    var uri = Uri.parse(packagePath);
-    var thisLibName = uri.pathSegments.first;
-    if (uri.path == from) {
-      return uri.pathSegments.last;
-    } else if (from.startsWith(thisLibName)) {
-      return p.posix.relative(uri.path, from: from).replaceFirst('../', '');
+  static String relative(String path, Uri to) {
+    var fileUri = Uri.parse(path);
+    var libName = to.pathSegments.first;
+    if ((to.scheme == 'package' &&
+            fileUri.scheme == 'package' &&
+            fileUri.pathSegments.first == libName) ||
+        (to.scheme == 'asset' && fileUri.scheme != 'package')) {
+      return p.posix
+          .relative(fileUri.path, from: to.path)
+          .replaceFirst('../', '');
     } else {
-      return packagePath;
+      return path;
     }
   }